// src/api/openapi/models/ReportedQuestion.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ReportedQuestion = {
    id?: string;
    userid?: string;
    forumid?: string;
    reason?: string;
    createdat?: string;
    updatedat?: string;
    question?: string;
    firstname?: string;
    lastname?: string;
    reported_username?: string;
    reported_firstname?: string;
    reported_userid?: string;
};